Integrating the swap mechanism to the frontend

+++ backend/db/db.sql
@@ -25,7 +25,7 @@ CREATE TABLE addresses (
   public_address uuid NOT NULL DEFAULT uuid_generate_v4(),
   user_id uuid DEFAULT uuid_generate_v4() NOT NULL REFERENCES users(id),
   coin_id INT NOT NULL REFERENCES coins(id),
-  balance INT NOT NULL,
+  balance NUMERIC NOT NULL,
   created_at DATE NOT NULL DEFAULT CURRENT_DATE
 );
