+++ backend/seeder.js
@@ -43,7 +43,7 @@ const importData = async () => {
       public_address uuid NOT NULL DEFAULT uuid_generate_v4(),
       user_id uuid DEFAULT uuid_generate_v4() NOT NULL REFERENCES users(id),
       coin_id INT NOT NULL REFERENCES coins(id),
-      balance INT NOT NULL,
+      balance NUMERIC NOT NULL,
       created_at DATE NOT NULL DEFAULT CURRENT_DATE
     );
     `);
